[ADD] Created CsvUtils static class and migrated LoadCSV method from Globals to the new CsvUtils class.
