+++ Assets/Scripts/Utilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a7d81cb4a31a884a8a74fb1aaf8f1c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
